scm: move git code into dvc/scm/git/... module
